Moved draw data to separate module and improved readability. Fixed bug relating to column offset.
