+++ src/main.rs
@@ -1,5 +1,6 @@
 #![cfg_attr(not(debug_assertions), windows_subsystem = "windows")] // hide console window on Windows in release
 mod config;
+mod draw;
 mod gui;
 
 fn main() -> Result<(), eframe::Error> {
